WIP: Unify storage access
This partially undoes the change done in commit 625d219f9e7ebf5f3e,
but it doesn't matter. The only reason I did that change was that
I needed to access the storage in @command_custom methods and
the 'instance' variable was not available there. There was no other
reason to change the interface to include the 'instance' parameter.
However now I actually need/really want the storage to be namespaced
by instance names.

Signed-off-by: Ondřej Lysoněk <olysonek@redhat.com>